requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
aiogram>=3.4
python-dotenv>=1.0
openai>=1.0.0